refactor: move help and aliases to command files
rel #2154
