+++ packages/pnpm/src/cmd/installTest.ts
@@ -1,8 +1,21 @@
+import renderHelp = require('render-help')
 import { PnpmOptions } from '../types'
-import install from './install'
-import { test } from './run'
+import { docsUrl } from './help'
+import { handler as install } from './install'
+import { handler as test } from './test'
 
-export default async function (input: string[], opts: PnpmOptions) {
+export const commandNames = ['install-test', 'it']
+
+export function help () {
+  return renderHelp({
+    aliases: ['it'],
+    description: 'Runs a \`pnpm install\` followed immediately by a \`pnpm test\`. It takes exactly the same arguments as \`pnpm install\`.',
+    url: docsUrl('install-test'),
+    usages: ['pnpm install-test'],
+  })
+}
+
+export async function handler (input: string[], opts: PnpmOptions) {
   await install(input, opts)
   await test(input, opts)
 }
