+++ packages/pnpm/src/cmd/prune.ts
@@ -1,9 +1,34 @@
+import renderHelp = require('render-help')
 import { InstallOptions, mutateModules } from 'supi'
 import createStoreController from '../createStoreController'
 import { readImporterManifestOnly } from '../readImporterManifest'
 import { PnpmOptions } from '../types'
+import { docsUrl, UNIVERSAL_OPTIONS } from './help'
 
-export default async (input: string[], opts: PnpmOptions) => {
+export const commandNames = ['prune']
+
+export function help () {
+  return renderHelp({
+    description: 'Removes extraneous packages',
+    descriptionLists: [
+      {
+        title: 'Options',
+
+        list: [
+          {
+            description: 'Remove the packages specified in \`devDependencies\`',
+            name: '--prod, --production',
+          },
+          ...UNIVERSAL_OPTIONS,
+        ],
+      },
+    ],
+    url: docsUrl('prune'),
+    usages: ['pnpm prune [--production]'],
+  })
+}
+
+export async function handler (input: string[], opts: PnpmOptions) {
   const store = await createStoreController(opts)
   return mutateModules([
     {
