+++ packages/pnpm/src/cmd/rebuild.ts
@@ -1,3 +1,5 @@
+import { oneLine } from 'common-tags'
+import renderHelp = require('render-help')
 import {
   rebuild,
   rebuildPkgs,
@@ -5,8 +7,41 @@ import {
 import createStoreController from '../createStoreController'
 import { readImporterManifestOnly } from '../readImporterManifest'
 import { PnpmOptions } from '../types'
+import { docsUrl, FILTERING, UNIVERSAL_OPTIONS } from './help'
 
-export default async function (
+export const commandNames = ['rebuild', 'rb']
+
+export function help () {
+  return renderHelp({
+    aliases: ['rb'],
+    description: 'Rebuild a package.',
+    descriptionLists: [
+      {
+        title: 'Options',
+
+        list: [
+          {
+            description: oneLine`Rebuild every package found in subdirectories
+              or every workspace package, when executed inside a workspace.
+              For options that may be used with \`-r\`, see "pnpm help recursive"`,
+            name: '--recursive',
+            shortAlias: '-r',
+          },
+          {
+            description: 'Rebuild packages that were not build during installation. Packages are not build when installing with the --ignore-scripts flag',
+            name: '--pending',
+          },
+          ...UNIVERSAL_OPTIONS,
+        ],
+      },
+      FILTERING,
+    ],
+    url: docsUrl('rebuild'),
+    usages: ['pnpm rebuild [<pkg> ...]'],
+  })
+}
+
+export async function handler (
   args: string[],
   opts: PnpmOptions,
   command: string,
