+++ packages/pnpm/src/cmd/restart.ts
@@ -1,0 +1,31 @@
+import renderHelp = require('render-help')
+import { handler as run } from './run'
+import { handler as start } from './start'
+import { handler as stop } from './stop'
+
+export const commandNames = ['restart']
+
+export function help () {
+  return renderHelp({
+    description: `Restarts a package. Runs a package's "stop", "restart", and "start" scripts, and associated pre- and post- scripts.`,
+    usages: ['pnpm restart [-- <args>...]'],
+  })
+}
+
+export async function handler (
+  args: string[],
+  opts: {
+    extraBinPaths: string[],
+    dir: string,
+    rawConfig: object,
+    argv: {
+      cooked: string[],
+      original: string[],
+      remain: string[],
+    },
+  },
+) {
+  await stop(args, opts)
+  await run(['restart', ...args], opts)
+  await start(args, opts)
+}
