+++ packages/pnpm/src/cmd/root.ts
@@ -1,6 +1,29 @@
 import path = require('path')
+import renderHelp = require('render-help')
 
-export default async function (
+export const commandNames = ['root']
+
+export function help () {
+  return renderHelp({
+    description: 'Print the effective \`node_modules\` directory.',
+    descriptionLists: [
+      {
+        title: 'Options',
+
+        list: [
+          {
+            description: 'Print the global \`node_modules\` directory',
+            name: '--global',
+            shortAlias: '-g',
+          },
+        ],
+      },
+    ],
+    usages: ['pnpm root [-g [--independent-leaves]]'],
+  })
+}
+
+export async function handler (
   args: string[],
   opts: {
     dir: string,
