+++ packages/pnpm/src/cmd/test.ts
@@ -1,0 +1,48 @@
+import { oneLine } from 'common-tags'
+import renderHelp = require('render-help')
+import { docsUrl, FILTERING } from './help'
+import { handler as run } from './run'
+
+export const commandNames = ['test', 't', 'tst']
+
+export function help () {
+  return renderHelp({
+    aliases: ['t', 'tst'],
+    description: `Runs a package's "test" script, if one was provided.`,
+    descriptionLists: [
+      {
+        title: 'Options',
+
+        list: [
+          {
+            description: oneLine`
+              Run the tests in every package found in subdirectories
+              or every workspace package, when executed inside a workspace.
+              For options that may be used with \`-r\`, see "pnpm help recursive"`,
+            name: '--recursive',
+            shortAlias: '-r',
+          },
+        ],
+      },
+      FILTERING,
+    ],
+    url: docsUrl('test'),
+    usages: ['pnpm test [-- <args>...]'],
+  })
+}
+
+export async function handler (
+  args: string[],
+  opts: {
+    extraBinPaths: string[],
+    dir: string,
+    rawConfig: object,
+    argv: {
+      cooked: string[],
+      original: string[],
+      remain: string[],
+    },
+  },
+) {
+  return run(['test', ...args], opts)
+}
