+++ packages/pnpm/src/cmd/unlink.ts
@@ -1,9 +1,43 @@
+import { oneLine } from 'common-tags'
+import renderHelp = require('render-help')
 import { mutateModules } from 'supi'
 import createStoreController from '../createStoreController'
 import { readImporterManifestOnly } from '../readImporterManifest'
 import { PnpmOptions } from '../types'
+import { docsUrl, UNIVERSAL_OPTIONS } from './help'
 
-export default async function (input: string[], opts: PnpmOptions) {
+export const commandNames = ['unlink', 'dislink']
+
+export function help () {
+  return renderHelp({
+    aliases: ['dislink'],
+    description: 'Removes the link created by \`pnpm link\` and reinstalls package if it is saved in \`package.json\`',
+    descriptionLists: [
+      {
+        title: 'Options',
+
+        list: [
+          {
+            description: oneLine`
+              Unlink in every package found in subdirectories
+              or in every workspace package, when executed inside a workspace.
+              For options that may be used with \`-r\`, see "pnpm help recursive"`,
+            name: '--recursive',
+            shortAlias: '-r',
+          },
+          ...UNIVERSAL_OPTIONS,
+        ],
+      },
+    ],
+    url: docsUrl('unlink'),
+    usages: [
+      'pnpm unlink (in package dir)',
+      'pnpm unlink <pkg>...',
+    ],
+  })
+}
+
+export async function handler (input: string[], opts: PnpmOptions) {
   const store = await createStoreController(opts)
   const unlinkOpts = Object.assign(opts, {
     storeController: store.ctrl,
